[WIN32K] Fix MSVC/x64 warnings

+++ win32ss/user/rtl/text.c
@@ -1146,7 +1146,7 @@ INT WINAPI DrawTextExWorker( HDC hdc,
     if (count == -1)
     {
 #ifdef _WIN32K_
-        count = wcslen(str);
+        count = (int)wcslen(str);
 #else
         count = strlenW(str);
 #endif
